feat(spatial): Phase B windows/doors + rooms UX polish
Openings: move drag (deferred pointer capture — 6px threshold distinguishes
tap→popover from drag→reposition/wall-change), server-first removeOpening.
registry: delete_room nulls light_positions.room_id before DELETE to avoid
FK violation; two new tests. spatial: f32/f64 cast fix.
rooms: hover pencil replaces rename button; collapse chevron with localStorage
persistence. style: collapse animation, pencil fade-in on hover.
roadmap: Phase A + B marked complete; Phase C Three.js spec added.

+++ coordinator/src/spatial.rs
@@ -103,7 +103,7 @@ impl SpatialEngine {
             let mut effective_azimuth = azimuth;
             let mut intensity_scale = 1.0;
 
-            let room_orientation = room.map(|r| r.orientation_degrees).unwrap_or(0.0);
+            let room_orientation = room.map(|r| r.orientation_degrees as f64).unwrap_or(0.0);
             if let Some(r) = room {
                 effective_azimuth = (azimuth - r.orientation_degrees as f64 + 360.0) % 360.0;
             }
@@ -120,7 +120,7 @@ impl SpatialEngine {
                 let mut contribution = 0.0_f32;
                 for o in room_openings {
                     let facing = wall_edge_to_degrees(&o.wall_edge);
-                    let adj = (facing - room_orientation as f64 + 360.0) % 360.0;
+                    let adj = (facing - room_orientation + 360.0) % 360.0;
                     let diff = (effective_azimuth - adj).abs();
                     let norm = if diff > 180.0 { 360.0 - diff } else { diff };
                     if norm <= 90.0 {
